integrate cn function for more straighforward styling
