+++ components/Tag.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 const Tag = ({ title }: { title: string }) => {
   return (
     <div className="rounded-sm bg-slate-light px-3 py-1 xl:px-[1.0625rem] xl:py-2.5">
