+++ components/homepage/AllDrinks.tsx
@@ -4,9 +4,9 @@ import { useEffect } from "react";
 import { useInView } from "react-intersection-observer";
 
 import { getFilteredDrinks } from "@/lib/actions/drinks.actions";
-import LoadingGraphic from "../graphics/LoadingGraphic";
 import { DrinkDataResponse, DrinksFilters } from "@/types/drinks.index";
 import useFetch from "@/hooks/useFetch";
+import LoadingGraphic from "../graphics/LoadingGraphic";
 import NoMoreDrinks from "../graphics/NoMoreDrinks";
 import DrinkCard from "../DrinkCard";
 import CustomButton from "../CustomButton";
